+++ frontend/src/pages/ProfilePage.js
@@ -1,6 +1,6 @@
 import Component from "../core/Component.js";
 import PageContainerWithLogo from "../components/UI/Container/PageContainerWithLogo.js";
-import Profile from "../components/UI/Profile/Profile.js";
+import Profile from "../components/Profile/Profile.js";
 import fetchAPI from "../utils/fetch/fetchAPI.js";
 
 export default class ProfilePage extends Component {
